Update and rename 2019.08.05.bio-lang.md to 2019.08.05.bio-decl.md
added some content to this post. Going to try and make it descriptive rather than just definitions, but for now I am building out a list and then will edit it into a descriptive post. Still working through the referenced papers and will start to define the "language of gene expression and scRNA-seq"